// src/types/product.ts
export interface ProductData {
  id: string;
  description: string;
  priceHT: number;
  priceTTC: number;
}

export interface ProductFormData {
  id: string;
  description: string;
  priceHT: string;
  priceTTC: string;
}

export interface LabelData {
  id: string;
  description: string;
  priceHT: number;
  priceTTC: number;
}